Add vote button and vote tracker to anecdotes
A vote button is now displayed to the left of the next anecdote button. The user can click the vote button an arbitrary number of times to add that many votes to the anecdote currently displayed. The vote counts will persist for each anecdote whether they're currently displayed or not